Movie browsing filters by Year are working.
I had to create a new field on the RDF DB with the year alone, because RDFLib does not support most of the Filters SPARQL supports.
It basically can only filter with the following operands: =, !=, <, >. To filter by year I had to create the new field...

You can test it if you go to Movies on the left sidebar of the site.

PS: the filtering table does not scale yet (I'll cap it later).